+++ Tests/MapboxMapsTests/Foundation/Extensions/Core/TransitionOptionsTests.swift
@@ -1,5 +1,5 @@
 import XCTest
-import MapboxMaps
+@testable import MapboxMaps
 
 final class TransitionOptionsTests: XCTestCase {
     func testInitWithNils() {
@@ -8,9 +8,9 @@ final class TransitionOptionsTests: XCTestCase {
             delay: nil,
             enablePlacementTransitions: nil)
 
-        XCTAssertNil(options.__duration)
-        XCTAssertNil(options.__delay)
-        XCTAssertNil(options.__enablePlacementTransitions)
+        XCTAssertNil(options.coreOptions.__duration)
+        XCTAssertNil(options.coreOptions.__delay)
+        XCTAssertNil(options.coreOptions.__enablePlacementTransitions)
     }
 
     func testInitWithNonNils() {
@@ -23,19 +23,18 @@ final class TransitionOptionsTests: XCTestCase {
             delay: delay,
             enablePlacementTransitions: enablePlacementTransition)
 
-        XCTAssertEqual(options.__duration, NSNumber(value: duration))
-        XCTAssertEqual(options.__delay, NSNumber(value: delay))
-        XCTAssertEqual(options.__enablePlacementTransitions, NSNumber(value: enablePlacementTransition))
+        XCTAssertEqual(options.coreOptions.__duration, NSNumber(value: duration))
+        XCTAssertEqual(options.coreOptions.__delay, NSNumber(value: delay))
+        XCTAssertEqual(options.coreOptions.__enablePlacementTransitions, NSNumber(value: enablePlacementTransition))
     }
 
     func testRefinedPropertiesWithNonNils() {
         let duration = TimeInterval.random(in: 0...10)
         let delay = TimeInterval.random(in: 0...10)
         let enablePlacementTransition = Bool.random()
-        let options = TransitionOptions(
-            __duration: NSNumber(value: duration),
-            delay: NSNumber(value: delay),
-            enablePlacementTransitions: NSNumber(value: enablePlacementTransition))
+        let options = TransitionOptions(duration: duration,
+                                        delay: delay,
+                                        enablePlacementTransitions: enablePlacementTransition)
 
         XCTAssertEqual(options.duration, duration)
         XCTAssertEqual(options.delay, delay)
@@ -44,12 +43,23 @@ final class TransitionOptionsTests: XCTestCase {
 
     func testRefinedPropertiesWithNils() {
         let options = TransitionOptions(
-            __duration: nil,
-            delay: nil,
-            enablePlacementTransitions: nil)
+                        duration: nil,
+                        delay: nil,
+                        enablePlacementTransitions: nil)
 
         XCTAssertNil(options.duration)
         XCTAssertNil(options.delay)
         XCTAssertNil(options.enablePlacementTransitions)
     }
+
+    func testTransitionOptionsEquality() {
+        let options = TransitionOptions(duration: 12, delay: 3, enablePlacementTransitions: true)
+        let options2 = TransitionOptions(duration: 12, delay: 3, enablePlacementTransitions: true)
+
+        XCTAssertEqual(options.duration, options2.duration)
+        XCTAssertEqual(options.delay, options2.delay)
+        XCTAssertEqual(options.enablePlacementTransitions, options2.enablePlacementTransitions)
+        XCTAssertEqual(options, options2)
+    }
+
 }